feat(condo): DOMA-11937 refactor phone page
