+++ apps/condo/domains/common/components/ResidentActions/ResidentActions.tsx
@@ -10,8 +10,8 @@ import { useLayoutContext } from '@condo/domains/common/components/LayoutContext
 import { MenuItem } from '@condo/domains/common/components/MenuItem'
 import { fontSizes } from '@condo/domains/common/constants/style'
 import { useSearchByPhoneModal } from '@condo/domains/common/hooks/useSearchByPhoneModal'
-import { searchByPhone } from '@condo/domains/contact/utils/clientCard'
 import { useTicketVisibility } from '@condo/domains/ticket/contexts/TicketVisibilityContext'
+import { searchByPhone } from '@condo/domains/ticket/utils/clientSchema/clientCard'
 
 import styles from './ResidentActions.module.css'
