+++ apps/condo/domains/contact/components/CreateContactForm.tsx
@@ -27,9 +27,9 @@ import { PhoneInput } from '@condo/domains/common/components/PhoneInput'
 import { useValidations } from '@condo/domains/common/hooks/useValidations'
 import { getObjectValueFromQuery } from '@condo/domains/common/utils/query'
 import { ContactRoleSelect } from '@condo/domains/contact/components/contactRoles/ContactRoleSelect'
-import { ClientCardTab, getClientCardTabKey } from '@condo/domains/contact/utils/clientCard'
 import { usePropertyValidations } from '@condo/domains/property/components/BasePropertyForm/usePropertyValidations'
 import { PropertyAddressSearchInput } from '@condo/domains/property/components/PropertyAddressSearchInput'
+import { ClientCardTab, getClientCardTabKey } from '@condo/domains/ticket/utils/clientSchema/clientCard'
 import { UnitNameInput, UnitNameInputOption } from '@condo/domains/user/components/UnitNameInput'
 import { UNABLE_TO_CREATE_CONTACT_DUPLICATE, UNABLE_TO_UPDATE_CONTACT_DUPLICATE } from '@condo/domains/user/constants/errors'
